# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TOPIDEALSPACE API",
        "description": "API REST de la marketplace TOPIDEALSPACE. Authentification via Laravel Sanctum (Bearer Token). Connectez-vous via **POST /auth/login**, copiez le token retourné, puis cliquez sur **Authorize** en haut à droite.",
        "contact": {
            "email": "contact@topidealspace.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local (dev)"
        }
    ],
    "paths": {
...
        "/admin/sub-categories/all": {
            "get": {
                "tags": [
                    "Admin Sous-Catégories"
                ],
                "summary": "Toutes les sous-catégories",
                "description": "Récupère toutes les sous-catégories groupées par catégorie",
                "operationId": "adminGetAllSubCategories",
                "responses": {
                    "200": {
                        "description": "Liste des sous-catégories par catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_categorie": {
                                                        "type": "integer"
                                                    },
                                                    "nom_categorie": {
                                                        "type": "string"
                                                    },
                                                    "sous_categories": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/categories/{categoryId}/sub-categories": {
            "get": {
                "tags": [
                    "Admin Sous-Catégories"
                ],
                "summary": "Sous-catégories d'une catégorie",
                "description": "Récupère toutes les sous-catégories d'une catégorie spécifique",
                "operationId": "adminGetSubCategoriesByCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des sous-catégories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_sous_categorie": {
                                                        "type": "integer"
                                                    },
                                                    "libelle_sous_categorie": {
                                                        "type": "string"
                                                    },
                                                    "produits_count": {
                                                        "type": "integer"
                                                    },
                                                    "can_delete": {
                                                        "type": "boolean"
                                                    },
                                                    "can_modify": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sub-categories/{id}": {
            "get": {
                "tags": [
                    "Admin Sous-Catégories"
                ],
                "summary": "Détail d'une sous-catégorie",
                "description": "Récupère une sous-catégorie avec ses produits associés",
                "operationId": "adminGetSubCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sous-catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la sous-catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SousCategorie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sous-catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Sous-Catégories"
                ],
                "summary": "Mettre à jour une sous-catégorie",
                "description": "Met à jour le libellé d'une sous-catégorie. Impossible si elle contient des produits.",
                "operationId": "adminUpdateSubCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sous-catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle_sous_categorie"
                                ],
                                "properties": {
                                    "libelle_sous_categorie": {
                                        "description": "Nouveau libellé",
                                        "type": "string",
                                        "maxLength": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sous-catégorie mise à jour"
                    },
                    "403": {
                        "description": "Sous-catégorie contient des produits"
                    },
                    "404": {
                        "description": "Sous-catégorie non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Sous-Catégories"
                ],
                "summary": "Supprimer une sous-catégorie",
                "description": "Supprime une sous-catégorie. Impossible si elle contient des produits.",
                "operationId": "adminDeleteSubCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sous-catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sous-catégorie supprimée avec succès"
                    },
                    "403": {
                        "description": "Sous-catégorie contient des produits"
                    },
                    "404": {
                        "description": "Sous-catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sub-categories": {
            "post": {
                "tags": [
                    "Admin Sous-Catégories"
                ],
                "summary": "Créer une sous-catégorie",
                "description": "Crée une nouvelle sous-catégorie dans une catégorie existante",
                "operationId": "adminCreateSubCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle_sous_categorie",
                                    "id_categorie"
                                ],
                                "properties": {
                                    "libelle_sous_categorie": {
                                        "description": "Libellé de la sous-catégorie",
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "id_categorie": {
                                        "description": "ID de la catégorie parente",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sous-catégorie créée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation ou sous-catégorie existante"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SousCategorie": {
                "properties": {
                    "id_sous_categorie": {
                        "type": "integer",
                        "example": 1
                    },
                    "libelle_sous_categorie": {
                        "type": "string",
                        "example": "Smartphones"
                    },
                    "id_categorie": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum Bearer Token. Entrez le token obtenu lors de la connexion.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}